change IngestResult numeric types to decimal

+++ precog/dotnet/PrecogClient/Client/Dto/IngestResult.cs
@@ -5,10 +5,10 @@ namespace Precog.Client.Dto
 	public class IngestResult
 	{
 		public bool Completed { get; private set; }
-		public float Total { get; set;}
-    	public float Ingested { get; set;}
-    	public float Failed { get; set;}
-    	public float Skipped { get; set;}
+		public decimal Total { get; set;}
+    	public decimal Ingested { get; set;}
+    	public decimal Failed { get; set;}
+    	public decimal Skipped { get; set;}
     	public string[] Errors { get; set;}
 
 		public IngestResult(bool completed= true)
